pyproject.toml
[project]
name = "xmrsigner"
version = "0.1.0"
description = "Build an offline, airgapped Monero signing device for less than $50!"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "embit>=0.8.0",
    "mock>=5.2.0",
    "monero @ git+https://github.com/DiosDelRayo/monero-python.git",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "polyseed @ git+https://github.com/DiosDelRayo/polyseed-python.git",
    "pytest>=8.4.1",
    "pyzbar @ git+https://github.com/seedsigner/pyzbar.git@c3c237821c6a20b17953efe59b90df0b514a1c03",
    "qrcode>=8.2",
    "qreader>=3.16",
    "urtypes @ git+https://github.com/jreesun/urtypes.git@e0d0db277ec2339650343eaf7b220fffb9233241",
]

# [project.optional-dependencies]
# hardware = [
#     "picamera==1.13",
#     "RPi.GPIO==0.7.0",
# ]
